% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada.R
\docType{data}
\name{canada}
\alias{canada}
\title{DATASET_TITLE}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{\code{indicator}}{character COLUMN_DESCRIPTION}
\item{\code{Value in 2018}}{double COLUMN_DESCRIPTION}
}
}
\usage{
canada
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
