% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_partners.R
\docType{data}
\name{canada_partners}
\alias{canada_partners}
\title{DATASET_TITLE}
\format{
A data frame with 5 rows and 4 variables:
\describe{
\item{\code{partner}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
\item{\code{share_imports}}{double COLUMN_DESCRIPTION}
\item{\code{epi}}{double COLUMN_DESCRIPTION}
}
}
\usage{
canada_partners
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
