% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_series.R
\docType{data}
\name{indicator_series}
\alias{indicator_series}
\title{DATASET_TITLE}
\format{
A data frame with 112640 rows and 7 variables:
\describe{
\item{\code{series}}{character COLUMN_DESCRIPTION}
\item{\code{iso}}{character COLUMN_DESCRIPTION}
\item{\code{country}}{character COLUMN_DESCRIPTION}
\item{\code{indicator}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{character COLUMN_DESCRIPTION}
\item{\code{value}}{double COLUMN_DESCRIPTION}
\item{\code{change}}{double COLUMN_DESCRIPTION}
}
}
\usage{
indicator_series
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
