% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_race.R
\docType{data}
\name{rat_race}
\alias{rat_race}
\title{DATASET_TITLE}
\format{
A data frame with 3470 rows and 6 variables:
\describe{
\item{\code{country_name}}{character COLUMN_DESCRIPTION}
\item{\code{iso}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
\item{\code{log_gdp_per_cap}}{double COLUMN_DESCRIPTION}
\item{\code{gini_disp}}{double COLUMN_DESCRIPTION}
\item{\code{hours}}{double COLUMN_DESCRIPTION}
}
}
\usage{
rat_race
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
